- fixed rounding bug in mimsdata, removed test button in stackediting
git-svn-id: https://zoomwhee.mgh.harvard.edu/svn_devel/OpenMIMS/trunk@208 ea379df5-90a0-4b1b-ac45-2d7255ffb434

+++ src/com/nrims/MimsData.java
@@ -247,7 +247,7 @@ public class MimsData extends javax.swing.JPanel {
             jLabel20.setText(tempstring);
             
             int raster = image.getRaster();
-            raster = java.lang.Math.round(raster/1000);
+            raster = java.lang.Math.round( ( (float)raster )/1000);
             tempstring = raster + " um";
             jLabel16.setText(tempstring);
